chore: Raspberry Pi: factor out boot partition adjustments
On Debian 11 "bullseye" (or derivatives) and earlier, the boot
partition is mounted to /uboot in the u-boot integration path.
As several scripts rely on the files located on the patition,
we need to adjust those.

To ease long term maintenance and sharing, this is factored out
to a common configuration file.

Changelog: Title
Ticket: None

Signed-off-by: TheYoctoJester <jester@theyoctojester.info>